+++ sdks/opik_optimizer/src/opik_optimizer/__init__.py
@@ -1,3 +1,6 @@
 import importlib.metadata
 
 __version__ = importlib.metadata.version("opik_optimizer")
+
+from .mipro_optimizer import MiproOptimizer
+from .few_shot_optimizer import FewShotOptimizer
