+++ sdks/opik_optimizer/src/opik_optimizer/integrations/dspy/__init__.py
@@ -1,0 +1,1 @@
+from .dspy_optimizer import DspyOptimizer
