--- sdks/opik_optimizer/src/opik_optimizer/utils.py
@@ -1,15 +1,0 @@
-class State(dict):
-    def __getattr__(self, key):
-        try:
-            return self[key]
-        except KeyError as e:
-            raise AttributeError(e)
-
-    def __setattr__(self, key, value):
-        self[key] = value
-
-    def __delattr__(self, key):
-        try:
-            del self[key]
-        except KeyError as e:
-            raise AttributeError(e)
